refactor(probe): extract shared target-service primitives module
